+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 1.0.2 (Next)
 
 * [#268](https://github.com/codegram/hyperclient/pull/268): Replace Travis-CI with GHA - [@dblock](https://github.com/dblock).
+* [#269](https://github.com/codegram/hyperclient/pull/269): Re-add code coverage - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ### 1.0.1 (2021/01/02)